+++ TenjinSDK.xcframework/ios-arm64_i386_x86_64-simulator/TenjinSDK.framework/PrivateHeaders/TenjinConst.h
@@ -1,0 +1,56 @@
+//endpoints
+#define kTenjinEventEndpoint @"event"
+#define kTenjinUserEndpoint @"user"
+#define kTenjinPurchaseEndpoint @"purchase"
+#define kTenjinAdRevenueEndpoint @"ad_impression"
+#define kTenjinAdMobAdRevenueEndpoint @"ad_impressions/admob"
+#define kTenjinAdRevenueAppLovinEndpoint @"ad_impressions/max"
+#define kTenjinAdRevenueIronSourceEndpoint @"ad_impressions/ironsource"
+#define kTenjinHyperBidRevenueEndpoint @"ad_impressions/hyperbid"
+#define kTenjinTopOnRevenueEndpoint @"ad_impressions/topon"
+#define kTenjinConversionEndpoint @"conversion-values"
+
+//this line replaced by build script
+#define kTenjinTenjinSDKVersion @"1.12.25"
+
+
+#define kTenjinPlatformIos @"ios"
+#define kTenjinHTTPMethodPOST @"POST"
+#define kTenjinHTTPMethodGET @"GET"
+#define kTenjinPrefsFile @"TenjinPreferences"
+
+#define kTenjinBundleId @"bundle_id"
+#define kTenjinAdvertisingId @"advertising_id"
+#define kTenjinVendorId @"developer_device_id"
+#define kTenjinPlatform @"platform"
+#define kTenjinAppVersion @"app_version"
+#define kTenjinAppSubversion @"app_subversion"
+#define kTenjinTrackingAuthorizationStatus @"tracking_status"
+#define kTenjinOsVersion @"os_version"
+#define kTenjinLimitAdTracking @"limit_ad_tracking"
+#define kTenjinSdkVersion @"sdk_version"
+#define kTenjinCountry @"country"
+#define kTenjinDevice @"device"
+#define kTenjinIad @"iad"
+#define kTenjinBuildId @"build_id"
+#define kTenjinOsVersionRelease @"os_version_release"
+#define kTenjinLocale @"locale"
+#define kTenjinDeviceInfo @"device_info"
+#define kTenjinDeviceModel @"device_model"
+#define kTenjinAAAttributionToken @"attribution_token"
+#define kTenjinCustomerUserId @"customer_user_id"
+
+#define kTenjinSessionId @"session_id"
+#define kTenjinSentAt @"sent_at"
+#define kTenjinSignature @"signature"
+#define kTenjinConversionValue @"conversion_value"
+#define kTenjinCoarseConversionValue @"coarse_conversion_value"
+#define kTenjinLockWindow @"lock_window"
+#define kTenjinRetryEnabled @"retry_enabled"
+
+//one hour in seconds
+#define kTenjinMaxWait 3600
+
+#define kTenjinStartupMaxRetries 5
+#define kTenjinRetrySleepBaseInSeconds 2.0
+#define kTenjinRetrySleepCapInSeconds 30.0